Adjusting vit models
